+++ blueocean-pipeline-api-impl/src/main/java/io/jenkins/blueocean/rest/impl/pipeline/BranchImpl.java
@@ -6,6 +6,8 @@ import hudson.Util;
 import hudson.model.BuildableItem;
 import hudson.model.Item;
 import hudson.model.Job;
+import io.jenkins.blueocean.pipeline.api.BlueMultiBranchPipeline.Branch;
+import io.jenkins.blueocean.pipeline.api.BlueMultiBranchPipeline.PullRequest;
 import io.jenkins.blueocean.rest.Navigable;
 import io.jenkins.blueocean.rest.Reachable;
 import io.jenkins.blueocean.rest.annotation.Capability;
@@ -15,14 +17,10 @@ import io.jenkins.blueocean.rest.model.BluePipeline;
 import io.jenkins.blueocean.rest.model.BluePipelineScm;
 import io.jenkins.blueocean.rest.model.Resource;
 import jenkins.branch.MultiBranchProject;
-import jenkins.scm.api.SCMHead;
-import jenkins.scm.api.metadata.ContributorMetadataAction;
 import jenkins.scm.api.metadata.ObjectMetadataAction;
 import jenkins.scm.api.metadata.PrimaryInstanceMetadataAction;
-import jenkins.scm.api.mixin.ChangeRequestSCMHead;
 import org.jenkinsci.plugins.workflow.job.WorkflowJob;
 import org.kohsuke.stapler.export.Exported;
-import org.kohsuke.stapler.export.ExportedBean;
 
 import static io.jenkins.blueocean.rest.model.KnownCapabilities.BLUE_BRANCH;
 import static io.jenkins.blueocean.rest.model.KnownCapabilities.JENKINS_WORKFLOW_JOB;
@@ -86,107 +84,6 @@ public class BranchImpl extends PipelineImpl {
         public Resource resolve(Item context, Reachable parent, Item target) {
             if (context==target.getParent()) {
                 return getPipeline(context,parent);
-            }
-            return null;
-        }
-    }
-
-    @ExportedBean
-    public static class Branch {
-
-        public static final String BRANCH = "branch";
-        private static final String BRANCH_URL = "url";
-        private static final String BRANCH_PRIMARY = "isPrimary";
-
-        private final String url;
-        private final boolean primary;
-
-        public Branch(String url, boolean primary) {
-            this.url = url;
-            this.primary = primary;
-        }
-
-        @Exported(name = BRANCH_URL)
-        public String getUrl() {
-            return url;
-        }
-
-        @Exported(name = BRANCH_PRIMARY)
-        public boolean isPrimary() {
-            return primary;
-        }
-
-        public static Branch getBranch(Job job) {
-            ObjectMetadataAction om = job.getAction(ObjectMetadataAction.class);
-            PrimaryInstanceMetadataAction pima = job.getAction(PrimaryInstanceMetadataAction.class);
-            if (om == null && pima == null) {
-                return null;
-            }
-            String url = om != null && om.getObjectUrl() != null ? om.getObjectUrl() : null;
-            return new Branch(url, pima != null);
-        }
-    }
-
-    @ExportedBean
-    public static class PullRequest {
-
-        public static final String PULL_REQUEST = "pullRequest";
-        private static final String PULL_REQUEST_NUMBER = "id";
-        private static final String PULL_REQUEST_AUTHOR = "author";
-        private static final String PULL_REQUEST_TITLE = "title";
-        private static final String PULL_REQUEST_URL = "url";
-
-        private final String id;
-
-        private final String url;
-
-        private final String title;
-
-        private final String author;
-
-        public PullRequest(String id, String url, String title, String author) {
-            this.id = id;
-            this.url = url;
-            this.title = title;
-            this.author = author;
-        }
-
-        @Exported(name = PULL_REQUEST_NUMBER)
-        public String getId() {
-            return id;
-        }
-
-
-        @Exported(name = PULL_REQUEST_URL)
-        public String getUrl() {
-            return url;
-        }
-
-
-        @Exported(name = PULL_REQUEST_TITLE)
-        public String getTitle() {
-            return title;
-        }
-
-
-        @Exported(name = PULL_REQUEST_AUTHOR)
-        public String getAuthor() {
-            return author;
-        }
-
-        public static PullRequest get(Job job) {
-            // TODO probably want to be using SCMHeadCategory instances to categorize them instead of hard-coding for PRs
-            SCMHead head = SCMHead.HeadByItem.findHead(job);
-            if(head instanceof ChangeRequestSCMHead) {
-                ChangeRequestSCMHead cr = (ChangeRequestSCMHead)head;
-                ObjectMetadataAction om = job.getAction(ObjectMetadataAction.class);
-                ContributorMetadataAction cm = job.getAction(ContributorMetadataAction.class);
-                return new PullRequest(
-                    cr.getId(),
-                    om != null ? om.getObjectUrl() : null,
-                    om != null ? om.getObjectDisplayName() : null,
-                    cm != null ? cm.getContributor() : null
-                );
             }
             return null;
         }
