+++ blueocean-pipeline-api-impl/src/main/java/io/jenkins/blueocean/rest/impl/pipeline/OrganizationFolderRunImpl.java
@@ -8,8 +8,6 @@ import io.jenkins.blueocean.rest.hal.Link;
 import io.jenkins.blueocean.rest.model.BlueActionProxy;
 import io.jenkins.blueocean.rest.model.BlueArtifactContainer;
 import io.jenkins.blueocean.rest.model.BlueChangeSetEntry;
-import io.jenkins.blueocean.rest.model.BluePipelineNodeContainer;
-import io.jenkins.blueocean.rest.model.BluePipelineStepContainer;
 import io.jenkins.blueocean.rest.model.BlueRun;
 import io.jenkins.blueocean.rest.model.BlueTestResultContainer;
 import io.jenkins.blueocean.rest.model.BlueTestSummary;
@@ -146,17 +144,7 @@ public class OrganizationFolderRunImpl extends BlueRun {
     }
 
     @Override
-    public BluePipelineNodeContainer getNodes() {
-        return null;
-    }
-
-    @Override
     public Collection<BlueActionProxy> getActions() {
-        return null;
-    }
-
-    @Override
-    public BluePipelineStepContainer getSteps() {
         return null;
     }
