+++ blueocean-pipeline-api/src/main/java/io/jenkins/blueocean/pipeline/api/BluePipelineNodeContainer.java
@@ -1,0 +1,9 @@
+package io.jenkins.blueocean.pipeline.api;
+
+import io.jenkins.blueocean.rest.model.Container;
+
+/**
+ * @author Vivek Pandey
+ */
+public abstract class BluePipelineNodeContainer extends Container<BluePipelineNode> {
+}
