--- blueocean-rest/src/main/java/io/jenkins/blueocean/rest/model/BluePipelineNodeContainer.java
@@ -1,7 +1,0 @@
-package io.jenkins.blueocean.rest.model;
-
-/**
- * @author Vivek Pandey
- */
-public abstract class BluePipelineNodeContainer extends Container<BluePipelineNode>{
-}
